Fix Dirty Flags
