+++ Simple Localizaion/Assets/DevsDaddy/SimpleLocalization/Core/Locales/LocaleTable.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: e73c1e202d264929b51071ed07f327f7
+timeCreated: 1711445896
