+++ Simple Localizaion/Assets/DevsDaddy/SimpleLocalization/Resources/Locales/Russian.asset
@@ -9,9 +9,18 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: e73c1e202d264929b51071ed07f327f7, type: 3}
   m_Name: Russian
-  m_EditorClassIdentifier: SimpleLocalization:DevsDaddy.SimpleLocalization.Core.Locales:LocaleTable
-  LocaleCode: 10
+  m_EditorClassIdentifier: 
+  LocaleCode: 30
   LocaleName: Russian
-  Strings: []
+  Strings:
+  - Code: WelcomeDesc
+    Value: "\u0421\u043F\u0430\u0441\u0438\u0431\u043E \u0437\u0430 \u0443\u0441\u0442\u0430\u043D\u043E\u0432\u043A\u0443
+      Simple Localization Toolkit!\n\u0412 \u044D\u0442\u043E\u043C \u043F\u0440\u0438\u043C\u0435\u0440\u0435
+      \u0432\u044B \u043C\u043E\u0436\u0435\u0442\u0435 \u043D\u0430\u0439\u0442\u0438
+      \u0440\u0430\u0431\u043E\u0442\u0443 \u0441 \u043A\u043E\u043C\u043F\u043E\u043D\u0435\u043D\u0442\u0430\u043C\u0438
+      \u0438 Fallback-\u0444\u0443\u043D\u043A\u0446\u0438\u044F\u043C\u0438!"
+  - Code: Welcome
+    Value: "\u0414\u043E\u0431\u0440\u043E \u043F\u043E\u0436\u0430\u043B\u043E\u0432\u0430\u0442\u044C
+      \u0432 Simple Localization!"
